CNDIT-1241 : Person Service : Implement Builder pattern and optimize to account for separating out the Organization Service. (#57)
* CNDIT-1241 :
Optimize the Person service to remove the organization service related logic
Update the Person Service to implement the Builder pattern
Update and verify all the unit tests.

* CNDIT-1241 :
Renaming the topics to be in consistent with the following convention :
https://cdc-nbs.atlassian.net/wiki/spaces/NM/pages/581664771/Reporting+services+naming+convention+for+all+components#Kafka-%2F-Debezium
Removing unnecessary kafka topics
Updating the logging timezone to be local.
Update the Kafka Producer to handle error conditions.

* CNDIT-1241 :
Addressing the code review comments.
Updating the name of couple of fields in the provider to follow the mapping doc
Migrated the json post processing call from the Model object to the DataProcessor utility class
Validated and verified all the unit tests.

* CNDIT-1241 :
Addressing the code review comments.
Migrating all the transformer logic to its own folder
Verified and validated all the uni tests

* CNDIT-1241 :
Minor update to the PersonService - Rename Transformer variable to be a relevant name.

* CNDIT-1241 :
Making the PersonTransformers Component and autowiring in the PersonService

* CNDIT-1241 :
Addressing the Code review comments. Processing the ProviderReporting `country` data to be  from the `code_short_desc_txt` in the provider stored proc.
Verified and validated all the unit tests.

* CNDIT-1241 :
Updating the Json field names in line with the stored proc for the post processing data fields.
Verified and validated all the unit tests.

* CNDIT-1241 :
Updating few more model mismatches. Adding the processing of the `within_city_limits` for the PatientReporting and PatientElastic data
Made model changes to the SSN processing.